# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.ProfessorCreateDTO:
    properties:
      formacao:
        type: string
      nome:
        type: string
    required:
    - formacao
    - nome
    type: object
  dtos.ProfessorUpdateDTO:
    properties:
      formacao:
        type: string
      nome:
        type: string
    type: object
  models.Aluno:
    properties:
      cpf:
        type: string
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cpf
    - nome
    type: object
  models.Curso:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      faculdade:
        $ref: '#/definitions/models.Faculdade'
      faculdade_unique_id:
        type: integer
      nome:
        type: string
      unique_id:
        type: integer
    type: object
  models.CursoDisciplina:
    properties:
      curso:
        $ref: '#/definitions/models.Curso'
      curso_unique_id:
        type: integer
      disciplina:
        $ref: '#/definitions/models.Disciplina'
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      disciplina_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.Disciplina:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      nome:
        type: string
      professor:
        $ref: '#/definitions/models.Professor'
      professor_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.DisciplinaMatricula:
    properties:
      aluno:
        $ref: '#/definitions/models.Aluno'
      aluno_unique_id:
        type: integer
      curso_disciplina:
        $ref: '#/definitions/models.CursoDisciplina'
      curso_disciplina_unique_id:
        type: integer
      data_matricula:
        type: string
      unique_id:
        type: integer
    type: object
  models.Faculdade:
    properties:
      cnpj:
        type: string
      cursos:
        items:
          $ref: '#/definitions/models.Curso'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cnpj
    - nome
    type: object
  models.Professor:
    properties:
      disciplinas:
        items:
          $ref: '#/definitions/models.Disciplina'
        type: array
      formacao:
        type: string
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - formacao
    - nome
    type: object
host: localhost:5000
info:
  contact:
    name: Tathiane Souza
  description: Esta é a documentação API em Golang
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API Faculdade passei direto
  version: "1.0"
paths:
...
  /professor:
    get:
      consumes:
      - application/json
      description: Busca todos os professors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Professor'
      summary: Buscar todos os professors
      tags:
      - Controller Professor
    post:
      consumes:
      - application/json
      description: 'Cria um professor, os campos que devem ser enviados são: nome
        e formacao'
      parameters:
      - description: professor
        in: body
        name: professor
        required: true
        schema:
          $ref: '#/definitions/dtos.ProfessorCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Professor'
      summary: Cria um professor
      tags:
      - Controller Professor
  /professor/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um professor
      parameters:
      - description: UniqueID do professor
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta um professor
      tags:
      - Controller Professor
    get:
      consumes:
      - application/json
      description: Busca professor por id
      parameters:
      - description: UniqueID do professor
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Professor'
      summary: Busca professor por id
      tags:
      - Controller Professor
    put:
      consumes:
      - application/json
      description: 'Edita um professor, os campos que podem ser enviados são: nome
        e formacao'
      parameters:
      - description: UniqueID do professor
        in: path
        name: id
        required: true
        type: integer
      - description: professor
        in: body
        name: professor
        required: true
        schema:
          $ref: '#/definitions/dtos.ProfessorUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Professor'
      summary: Edita um professor
      tags:
      - Controller Professor
swagger: "2.0"